remove verbose WHR logging, the errors were caused by the SpringBattlePlayers table missing. recount active players before calculating whr percentile.

+++ ZkData/Ef/WHR/Ranks.cs
@@ -100,7 +100,7 @@ namespace Ratings
             var rating = acc.AccountRatings.Where(x => x.RatingCategory == cat).FirstOrDefault();
             var ladderElo = rating?.LadderElo ?? WholeHistoryRating.DefaultRating.LadderElo;
             if (!allowLoss && !allowGain) return (float)ladderElo;
-            if (double.IsNaN(targetRating))
+            if (float.IsNaN(targetRating))
             {
                 Trace.TraceWarning("Target rating for player " + acc.Name + "(" + acc.AccountID + ") is NaN");
                 return (float)ladderElo;
